Remove `GeometryProvider` in favor of spawn hooks
