+++ readme.md
@@ -67,7 +67,7 @@ use bevy_trenchbroom::class::builtin::*;
 // If you're using a BSP workflow, this base class includes a bunch
 // of useful compiler properties.
 #[base(BspWorldspawn)]
-#[geometry(GeometryProvider::new().trimesh_collider().smooth_by_default_angle().with_lightmaps())]
+#[spawn_hooks(SpawnHooks::new().trimesh_collider())]
 pub struct Worldspawn {
     /// A useful example property.
     pub fog_color: Color,
@@ -89,7 +89,7 @@ pub struct MyBaseClass {
 #[derive(SolidClass, Component, Reflect)]
 #[reflect(QuakeClass, Component)]
 #[base(Visibility, MyBaseClass)]
-#[geometry(GeometryProvider::new().trimesh_collider().smooth_by_default_angle().with_lightmaps())]
+#[spawn_hooks(SpawnHooks::new().trimesh_collider())]
 // By default, names are converted into snake_case.
 // Using the classname attribute, you can define the case you want
 // it to be converted to instead.
@@ -103,8 +103,7 @@ pub struct FuncWall;
 // If you're using a BSP workflow, this base class includes a bunch
 // of useful compiler properties.
 #[base(BspSolidEntity)]
-// Don't include a collider for func_illusionary.
-#[geometry(GeometryProvider::new().smooth_by_default_angle().with_lightmaps())]
+// Don't include spawn hooks adding a collider for func_illusionary.
 pub struct FuncIllusionary;
 
 // A more advanced example
@@ -280,7 +279,7 @@ Currently, PVS data generated by `vis` isn't used.
 
 `bevy_trenchbroom` supports [bevy_rapier3d](https://crates.io/crates/bevy_rapier3d) and [avian3d](https://crates.io/crates/avian3d) to easily add colliders when spawning geometry.
 
-First, enable the `rapier` or `avian` feature on the crate, then either call `convex_collider` or `trimesh_collider` on your class's `GeometryProvider` to create the respective type of collider(s) with said geometry.
+First, enable the `rapier` or `avian` feature on the crate, then either call `convex_collider` or `trimesh_collider` on your class's `SpawnHooks` to create the respective type of collider(s) with said geometry.
 
 ## Multiplayer
